% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvConversion.R
\name{cleanForCSV}
\alias{cleanForCSV}
\title{Helper function for non-string metadata in rmmToCSV}
\usage{
cleanForCSV(x = NULL)
}
\arguments{
\item{x}{An \code{rmm} entry that will be entered into an \code{rmmToCSV} function.}
}
\description{
Cleans up metadata instances that get messy if one tries to write them directly to csv tables (i.e. extent objects, bibtex objects.)
}
\details{
See Examples.
}
\examples{
rmm=rmmTemplate()
rmm=rmmAutofillPackageCitation(rmm,c('raster','sp'))
cleanForCSV(rmm$data$environment$extent)

}
\author{
Hannah Owens <hannah.owens@gmail.com>, Cory Merow <cory.merow@gmail.com>
}
