% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvToRMM.R
\name{csvToRMM}
\alias{csvToRMM}
\title{Recreate rangeModelMetaData Object From .csv File}
\usage{
csvToRMM(csv)
}
\arguments{
\item{csv}{A character file path to the csv file.}
}
\description{
Takes user-input .csv file and converts it back to a rangeModelMetaData object.
}
\details{
See Examples.
}
\examples{

}
\author{
Jamie Kass <jamie.m.kass@gmail.com>
}
