% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTemplate.R
\name{rangeModelMetadataTemplate}
\alias{rangeModelMetadataTemplate}
\title{Range modeling metadata}
\usage{
rangeModelMetadataTemplate(useCase = "apAll")
}
\arguments{
\item{useCase}{character string; 'apAll', 'apObligate', 'apMinimal'}
}
\value{
a range model metadata list
}
\description{
Make an empty metadata list
}
\details{
See Examples.
}
\examples{
rmm1=rangeModelMetadataTemplate()
rmm2=rangeModelMetadataTemplate(useCase="apObligate")
str(rmm2)

}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>
}
