% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillENMeval}
\alias{rmmAutofillENMeval}
\title{Fill in relevant rmm fields from an ENMevaluation object.}
\usage{
rmmAutofillENMeval(rmm, ENMevaluation)
}
\arguments{
\item{ENMevaluation}{an ENMevaluation object}

\item{rrm}{an rmm list}
}
\description{

}
\details{
See Examples.
}
\author{
Jamie M. Kass <jamie.m.kass@gmail.com>
}

