% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillEnvironment}
\alias{rmmAutofillEnvironment}
\title{Add relevant environmental data information to an rmm object}
\usage{
rmmAutofillEnvironment(rmm, env, transfer = FALSE)
}
\arguments{
\item{env}{a raster stack}

\item{rrm}{an rmm list}
}
\description{
This can be used with environmental layers used for fitting or transferring
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

