% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillPackageCitation}
\alias{rmmAutofillPackageCitation}
\title{Add all package citations to an rmm object}
\usage{
rmmAutofillPackageCitation(rmm, packages)
}
\arguments{
\item{rmm}{an rmm list}

\item{packages}{a vector of quoted package names}
}
\description{
Using bibtex citations
}
\details{
See Examples.
}
\examples{
rmm=rangeModelMetadataTemplate(useCase='apAll')
rmm=rmmAutofillPackageCitation(rmm,c('raster','sp'))

}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

