% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillPrediction}
\alias{rmmAutofillPrediction}
\title{Add relevant model prediction info to an rmm object}
\usage{
rmmAutofillPrediction(rmm, prediction)
}
\arguments{
\item{prediction}{a raster layer or stack}

\item{rrm}{an rmm list}
}
\description{

}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

