% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillspocc}
\alias{rmmAutofillspocc}
\title{Add occurrence metadata from a spocc query to an rmm object}
\usage{
rmmAutofillspocc(rmm, occ)
}
\arguments{
\item{rmm}{an rmm list}

\item{occ}{Output from \code{\link[spocc]{occ}}}
}
\value{
a range model metadata list
}
\description{
This function populates occurrence field in an rmm object with output from a spocc query
}
\details{
See Examples.
}
\examples{
rmm=rmmTemplate()
xs <- spocc::occ(species="Xanthium strumarium")
rmmAutofillspocc(rmm = rmm, occ = xs)

}
\seealso{
\code{\link[spocc]{occ}}

Other autofill: \code{\link{rmmAutofillBIEN}},
  \code{\link{rmmAutofillENMeval}},
  \code{\link{rmmAutofillEnvironment}},
  \code{\link{rmmAutofillModelObj}},
  \code{\link{rmmAutofillPackageCitation}},
  \code{\link{rmmAutofillPrediction}}
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
