% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{rmmCheckMissingNames}
\alias{rmmCheckMissingNames}
\title{Check for missing fields}
\usage{
rmmCheckMissingNames(rmm, useCase = "apObligate")
}
\arguments{
\item{rmm}{a range model metadata list}

\item{useCase}{The rmm useCase to check the rmm against}
}
\value{
A vector of names that are missing from the rmm object.
}
\description{
Identify obligate fields that are missing
}
\details{
See Examples.
}
\examples{
rmm<-rangeModelMetadataTemplate() # Make an empty template


}
\seealso{
Other check: \code{\link{rmmCheckFinalize}},
  \code{\link{rmmCheckName}}, \code{\link{rmmCheckValue}}
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
