% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmmCheckShiny.R
\name{rmmCheckShiny}
\alias{rmmCheckShiny}
\title{RangeModelMetadata Check in Shiny}
\usage{
rmmCheckShiny()
}
\value{
None
}
\description{
Run shiny app to visualize rmm check functions
}
\details{
See Examples.
}
\note{
This function launches a shiny app in the default web browser
}
\examples{
rmm1=rangeModelMetadataTemplate(useCase='apAll')
rmm1=rmmAutofillPackageCitation(rmm1,c('raster','sp'))
r.f=system.file("extdata/Env_Demo",package='rangeModelMetadata')
raster.files=list.files(r.f,full.names = TRUE)
env=raster::stack(raster.files)
# for fitting environment
rmm1=rmmAutofillEnvironment(rmm1,env,transfer=0)
# for transfer environment 1 (assuming different than for fitting)
rmm1=rmmAutofillEnvironment(rmm1,env,transfer=1)
# for transfer environment 2 (assuming different than 1)
rmm1=rmmAutofillEnvironment(rmm1,env,transfer=2)
\dontrun{ rmmCheckShiny(rmm1) }

}
\author{
Jamie Kass <jamie.m.kass@gmail.com>
}
