% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{rmmCleanEmpties}
\alias{rmmCleanEmpties}
\title{Clean unused fields from range model metadata list}
\usage{
rmmCleanEmpties(rmm)
}
\arguments{
\item{rmm}{a range model metadata list}
}
\description{
Only optional fields that are empty are removed.
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

