% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{rmmCleanNULLs}
\alias{rmmCleanNULLs}
\title{Remove NULL entries range model metadata list}
\usage{
rmmCleanNULLs(rmm)
}
\arguments{
\item{rmm}{a range model metadata list}
}
\value{
printout to the console
}
\description{
Check if fields are NULL in a range model metadata list and toss
}
\details{
See Examples.
}
\examples{
# see vignette('rmm_vignette')
}
\seealso{
Other check: \code{\link{rmmCheckEmpty}},
  \code{\link{rmmCheckFinalize}},
  \code{\link{rmmCheckMissingNames}},
  \code{\link{rmmCheckName}}, \code{\link{rmmCheckValue}}
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
\concept{check}
