% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmmFamilies}
\alias{rmmFamilies}
\title{Print supported family names for rmm objects}
\usage{
rmmFamilies()
}
\value{
a vector of characters, indicating which RMMS families are supported
}
\description{
Used to see options to for specifying an rmm object template
}
\examples{
rmmFamilies()
}
