% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests.R
\name{rmmSuggest}
\alias{rmmSuggest}
\title{Suggest inputs for a range model metadata list}
\usage{
rmmSuggest(charString)
}
\description{
Supply fields to receive suggested inputs
}
\details{
See Examples.
}
\note{
strings should note have a leading $ (fix this to remove it!)
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

