% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTemplate.R
\name{rmmTemplate}
\alias{rmmTemplate}
\title{Range modeling metadata}
\usage{
rmmTemplate(useCase = "apAll")
}
\arguments{
\item{useCase}{character string; specifies an application profile (use case) that has been predefined. Currently supported are: 'apAll', 'apObligate', 'apMinimal'. Superceeds}

\item{families}{character vector; an alternative to specifying `useCase`. Provide a vector of family names to include all entities in a family in the template. Use `rmmFamilyNames` to see supported values.}
}
\value{
a range model metadata list
}
\description{
Make an empty metadata list
}
\details{
See Examples.
}
\examples{
rmm1=rmmTemplate()
rmm2=rmmTemplate(useCase="apObligate")
str(rmm2)

}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>
}
