% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvConversion.R
\name{cleanForCSV}
\alias{cleanForCSV}
\title{Helper function for non-string metadata in rmmToCSV}
\usage{
cleanForCSV(x = NULL)
}
\arguments{
\item{x}{An \code{rmm} entry that returned to the \code{rmmToCSV} function.}
}
\value{
Reformatted element for use in \code{rmmToCSV} function.
}
\description{
Cleans up metadata instances that get messy if one tries to write them directly to csv tables (i.e. extent objects, bibtex objects.)
}
\details{
This is a utility function for use by \code{rmmToCSV}.
}
\seealso{
Other csvConversion: \code{\link{csvToRMM}},
  \code{\link{rmmToCSV}}
}
\author{
Hannah Owens <hannah.owens@gmail.com>, Cory Merow <cory.merow@gmail.com>
}
\concept{csvConversion}
