% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvConversion.R
\name{csvToRMM}
\alias{csvToRMM}
\title{Recreate rangeModelMetaData Object From .csv File}
\usage{
csvToRMM(csv, useCase = NULL)
}
\arguments{
\item{csv}{A character file path to the csv file.}

\item{useCase}{character string; specifies an application profile (use case) by specifiying the families of entitiies that should be included. Specifying NULL includes all entities. Use `rmmFamilyNames` to see supported values.}
}
\description{
Takes user-input .csv file and converts it back to a rangeModelMetaData object.
}
\details{
See Examples.
}
\author{
Hannah Owens <hannah.owens@gmail.com>
}
