% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvConversion.R
\name{csvToRMM}
\alias{csvToRMM}
\title{Create rangeModelMetaData (`rmm`) object from a .csv File}
\usage{
csvToRMM(csv, family = NULL)
}
\arguments{
\item{csv}{A character file path to the csv file.}

\item{family}{character string; specifies an application profile (use case) by specifiying the families of entitiies that should be included. Specifying NULL includes all entities. Use \code{rmmFamilies()} to see supported values.}
}
\value{
An \code{rmm} object that was read from the supplied .csv text file.
}
\description{
Takes user-input .csv file and converts it to a rangeModelMetaData (`rmm`) object.
}
\details{
See Examples.
}
\examples{
csv <- "somePathOnYourMachine/rmm_example.csv";
\dontrun{temp <- csvToRMM(csv);}

}
\seealso{
Other csvConversion: \code{\link{cleanForCSV}},
  \code{\link{rmmToCSV}}
}
\author{
Hannah Owens <hannah.owens@gmail.com>
}
\concept{csvConversion}
