% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillENMeval}
\alias{rmmAutofillENMeval}
\title{Fill in relevant rmm fields from an ENMevaluation object.}
\usage{
rmmAutofillENMeval(rmm, e)
}
\arguments{
\item{rrm}{an rmm list}

\item{ENMevaluation}{an ENMevaluation object}
}
\value{
a range model metadata list
}
\description{
Fill in relevant rmm fields from an ENMevaluation object.
}
\details{
See Examples.
}
\author{
Jamie M. Kass <jamie.m.kass@gmail.com>
}

