% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillENMeval}
\alias{rmmAutofillENMeval}
\title{Fill in relevant rmm fields from an ENMevaluation object.}
\usage{
rmmAutofillENMeval(rmm, e, i)
}
\arguments{
\item{e}{an ENMevaluation object}

\item{i}{a numeric index value referring to the chosen model
(e.g., if you chose the model corresponding to row 5 in the results table, this number would be 5)}

\item{rrm}{an rmm list}
}
\value{
a range model metadata list
}
\description{
Fill in relevant rmm fields from an ENMevaluation object.
}
\details{
See Examples.
}
\author{
Jamie M. Kass <jamie.m.kass@gmail.com>
}

