% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillENMeval}
\alias{rmmAutofillENMeval}
\title{Fill in relevant rmm fields from an ENMevaluation object.}
\usage{
rmmAutofillENMeval(rmm, e, selectionCriteria, optimalModelIndex)
}
\arguments{
\item{rmm}{an rmm list}

\item{e}{an ENMevaluation object}

\item{selectionCriteria}{a character string indicating the model selection
rules used (e.g., "first chose models with lowest MTP omission rate, then
chose the model with highest average test AUC")}

\item{optimalModelIndex}{a numeric value indicating the row number of the
model chosen by the user
(e.g., if you chose the model corresponding to row 5 in the results table,
this number would be 5); multiple models
may be selected in theory (for the purposes of model averaging, etc.), but
selecting one is preferable to reduce confusion}
}
\value{
a range model metadata list
}
\description{
Fill in relevant rmm fields from an ENMevaluation object.
}
\details{
See Examples.
}
\examples{
#see vignette('rmm_workflow')

}
\seealso{
Other autofill: \code{\link{rmmAutofillBIEN}},
  \code{\link{rmmAutofillEnvironment}},
  \code{\link{rmmAutofillPackageCitation}},
  \code{\link{rmmAutofillspocc}}
}
\author{
Jamie M. Kass <jamie.m.kass@gmail.com>
}
\concept{autofill}
