% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillEnvironment}
\alias{rmmAutofillEnvironment}
\title{Add relevant environmental data information to an rmm object}
\usage{
rmmAutofillEnvironment(rmm, env, transfer)
}
\arguments{
\item{rmm}{an rmm list}

\item{env}{a raster stack}

\item{transfer}{0 if not transfer, 1:n for n environments that you're transferring to

raster.files=list.files(system.file("extdata/Env_Demo",package='rangeModelMetadata'),full.names = T)
env=raster::stack(raster.files)
# for fitting environment
# for the first environment that you're transfering to
# for the second environment that you're transfering to, etc.}
}
\description{
This can be used with environmental layers used for fitting or transferring
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

