% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillModelObj}
\alias{rmmAutofillModelObj}
\title{Add relevant model info to an rmm object}
\usage{
rmmAutofillModelObj(rmm, modelObj)
}
\arguments{
\item{rrm}{an rmm list}
}
\description{

}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

