% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillModelObj}
\alias{rmmAutofillModelObj}
\title{Add relevant model info to an rmm object}
\usage{
rmmAutofillModelObj(rmm, modelObj)
}
\arguments{
\item{rmm}{an rmm list}

\item{modelObj}{a model object}
}
\description{
Does stuff
}
\details{
See Examples.
}
\seealso{
Other autofill: \code{\link{rmmAutofillBIEN}},
  \code{\link{rmmAutofillENMeval}},
  \code{\link{rmmAutofillEnvironment}},
  \code{\link{rmmAutofillPackageCitation}},
  \code{\link{rmmAutofillPrediction}},
  \code{\link{rmmAutofillspocc}}
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
