% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillPackageCitation}
\alias{rmmAutofillPackageCitation}
\title{Add all package citations to an rmm object}
\usage{
rmmAutofillPackageCitation(packages)
}
\description{

}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

