% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillPackageCitation}
\alias{rmmAutofillPackageCitation}
\title{Add all package citations to an rmm object}
\usage{
rmmAutofillPackageCitation(rmm, packages)
}
\arguments{
\item{rmm}{an rmm list}

\item{packages}{a vector of quoted package names}
}
\value{
a range model metadata list
}
\description{
Using bibtex citations
}
\details{
See Examples.
}
\examples{
rmm=rmmTemplate()
rmm=rmmAutofillPackageCitation(rmm,c('raster','sp'))

}
\seealso{
Other autofill: \code{\link{rmmAutofillBIEN}},
  \code{\link{rmmAutofillENMeval}},
  \code{\link{rmmAutofillEnvironment}},
  \code{\link{rmmAutofillspocc}}
}
\author{
Brian Maitner <bmaitner@gmail.com>, Cory Merow <cory.merow@gmail.com>
}
\concept{autofill}
