% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill.R
\name{rmmAutofillPrediction}
\alias{rmmAutofillPrediction}
\title{Add relevant model prediction info to an rmm object}
\usage{
rmmAutofillPrediction(rmm, prediction)
}
\arguments{
\item{rmm}{an rmm list}

\item{prediction}{a raster layer or stack}
}
\value{
a range model metadata list
}
\description{
Add relevant model prediction info to an rmm object
}
\details{
See Examples.
}
\seealso{
Other autofill: \code{\link{rmmAutofillBIEN}},
  \code{\link{rmmAutofillENMeval}},
  \code{\link{rmmAutofillEnvironment}},
  \code{\link{rmmAutofillModelObj}},
  \code{\link{rmmAutofillPackageCitation}},
  \code{\link{rmmAutofillspocc}}
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
