% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmmCheckShiny.R
\name{rmmCheckShiny}
\alias{rmmCheckShiny}
\title{RangeModelMetadata Check in Shiny}
\usage{
rmmCheckShiny()
}
\value{
None
}
\description{
Run shiny app to visualize rmm check functions
}
\details{
See Examples.
}
\note{
This function launches a shiny app in the default web browser
}
\examples{
rmmCheckShiny(rmm)

}
\author{
Jamie Kass <jamie.m.kass@gmail.com>
}
