% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmmCheckShiny.R
\name{rmmCheckShiny}
\alias{rmmCheckShiny}
\title{RangeModelMetadata Check in Shiny}
\usage{
rmmCheckShiny()
}
\value{
None
}
\description{
Run shiny app to visualize rmm check functions
}
\details{
See Examples.
}
\note{
This function launches a shiny app in the default web browser
}
\examples{
\dontrun{
rmm1=rmmTemplate()
rmm1=rmmAutofillPackageCitation(rmm1,c('raster','sp'))
rasterFiles=list.files(path=paste(system.file(package='dismo'), '/ex', sep=''),
                       pattern='grd', full.names=TRUE)
make a stack of the rasters
env=raster::stack(rasterFiles)
# for fitting environment
rmm1=rmmAutofillEnvironment(rmm1,env,transfer=0)
# for transfer environment 1 (assuming different than for fitting)
rmm1=rmmAutofillEnvironment(rmm1,env,transfer=1)
# for transfer environment 2 (assuming different than 1)
rmm1=rmmAutofillEnvironment(rmm1,env,transfer=2)
}
\dontrun{ rmmCheckShiny(rmm1) }

}
\author{
Jamie Kass <jamie.m.kass@gmail.com>
}
