% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmmDataDictionary}
\alias{rmmDataDictionary}
\title{Open range model metadata dictionary.}
\usage{
rmmDataDictionary(excel = F)
}
\arguments{
\item{excel}{logical; open in excel?}
}
\description{
For viewing only
}
\examples{
dd=rmmDataDictionary()
}
