% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{rmmNameCheck}
\alias{rmmNameCheck}
\title{Check field names of a range model metadata list against convections}
\usage{
rmmNameCheck(rmm)
}
\arguments{
\item{rmm}{a range model metadata list}
}
\description{
Identify nonstandard fields
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}

