% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empties.R
\name{rmmPrintEmpty}
\alias{rmmPrintEmpty}
\title{Check if fields are empty in a range model metadata list}
\usage{
rmmPrintEmpty(rmm, obligateOnly = FALSE)
}
\arguments{
\item{rmm}{a range model metadata list}

\item{obligateOnly}{logical; only show empty obligat fields}
}
\value{
a range model metadata list
}
\description{
Check if fields are empty in a range model metadata list
}
\details{
See Examples.
}
\examples{
rmm=rangeModelMetadataTemplate(useCase='apAll')
r.f=system.file("extdata/Env_Demo",package='rangeModelMetadata')
raster.files=list.files(r.f,full.names = TRUE)
env=raster::stack(raster.files)
# for fitting environment
rmm=rmmAutofillEnvironment(rmm,env,transfer=0)
# for the first environment that you're transfering to
rmm=rmmAutofillEnvironment(rmm,env,transfer=1)
# for the second environment that you're transfering to, etc.
rmm=rmmAutofillEnvironment(rmm,env,transfer=2)

}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
