% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests.R
\name{rmmSuggest}
\alias{rmmSuggest}
\title{Suggest inputs for a range model metadata list}
\usage{
rmmSuggest(charString, fullFieldDepth = FALSE)
}
\arguments{
\item{charString}{string referencing fields of the form `field1$field2` or `field1$field2$field3`, etc.}

\item{fullFieldDepth}{print all fields below the current field depth

rmm1=rmmTemplate()
rmmSuggest('dataPrep',fullFieldDepth=FALSE)
rmmSuggest('dataPrep',fullFieldDepth=TRUE)
rmmSuggest('dataPrep$errors$duplicateRemoval')
rmmSuggest('dataPrep$errors$duplicateRemoval$rule')
rmmSuggest('model')
rmmSuggest('modelFit$algorithmSettings$')
rmmSuggest('modelFit$algorithmSettings$maxent$')
rmmSuggest('$modelFit$algorithmSettings$maxent$featureSet')}
}
\description{
Supply fields to receive suggested inputs
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>,
}
