% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataTemplate.R
\name{rmmTemplate}
\alias{rmmTemplate}
\title{Range modeling metadata}
\usage{
rmmTemplate(family = NULL)
}
\arguments{
\item{family}{character string; specifies an application profile (use case) by specifiying the families of entitiies that should be included. Specifying NULL includes all entities. Use `rmmFamilies` to see supported values.}
}
\value{
a range model metadata list
}
\description{
Make an empty metadata list
}
\details{
See Examples.
}
\examples{
rmm1=rmmTemplate()
rmm2=rmmTemplate(family=c('base'))
str(rmm2)

}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>
}
