% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCSV.R
\name{rmmToCSV}
\alias{rmmToCSV}
\title{Create .csv File From rangeModelMetaData Object}
\usage{
rmmToCSV(x = rangeModelMetadataTemplate("apAll"), filename = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{rmm}} that the user wishes transposed into a .csv file.}

\item{filename}{The name of the transcription .csv file.}
}
\description{
Takes user-input rangeModelMetaData object and from it generates a .csv file that can be used to document range model metadata for a variety of applications.
}
\details{
See Examples.
}
\examples{
outFile=system.file("extdata/demo_rmmToCSV.csv",package='rangeModelMetadata')
rmmObj=rangeModelMetadataTemplate('apAll')
#rmmToCSV(rmmObj,filename=outFile)
#system(paste0('open ' outFile))
}
\author{
Cory Merow <cory.merow@gmail.com>, Brian Maitner <bmaitner@gmail.com>, Hannah Owens <hannah.owens@gmail.com
}

