% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmmToCSV.R
\name{rmmToCSV}
\alias{rmmToCSV}
\title{Create .csv File From rangeModelMetaData Object}
\usage{
rmmToCSV(x = rangeModelMetadataTemplate(useCase = "apAll"), filename = NULL)
}
\arguments{
\item{x}{An object of class \code{rmm} that the user wishes transposed into a .csv file.}

\item{filename}{The name of the transcription .csv file.}
}
\description{
Takes user-input rangeModelMetaData object and from it generates a .csv file that can be used to document range model metadata for a variety of applications.
}
\details{
See Examples.
}
\examples{
rmm=rangeModelMetadataTemplate(useCase='apAll')
r.f=system.file("extdata/Env_Demo",package='rangeModelMetadata')
raster.files=list.files(r.f,full.names = TRUE)
env=raster::stack(raster.files)
# for fitting environment
rmm=rmmAutofillEnvironment(rmm,env,transfer=0)
# for the first environment that you're transfering to
rmm=rmmAutofillEnvironment(rmm,env,transfer=1)
# for the second environment that you're transfering to, etc.
rmm=rmmAutofillEnvironment(rmm,env,transfer=2)
\dontrun{
rmmToCSV(rmm,file='somePathOnYourMachine/rmm_example.csv')
}
}
\author{
Hannah Owens <hannah.owens@gmail.com>, Cory Merow <cory.merow@gmail.com>
}
